Add focus to hover rules
This commit will remove the outline when tabbing through for some parts
of the website. Instead of a strong outline around some links that wrap
images or headers, the hover state (which is easily visible) will be
shown instead.

In places where the change from normal to hover is not as apparent, the
outline will remain.