Add theme styles to all project components
Configure all components to use css variables for many of their colors.
These variables get modified whenever the body tag gets a different
class applied to it. Either:
  - light
  - dark
  - solarized

The home page will set it's background images to match the current
theme. If a theme is selected while still on the home page, the
theme service will emit an item to its subject. This subject is
subscribed to in the home.component.ts file which will add and
remove the necessary classes on the top level home div.